# swagger/swagger.yaml
swagger: '2.0'
info:
  description: Swagger API for SmartbeeExam1.
  version: 1.0.0
  title: SmartbeeExam1
host: localhost
basePath: /v1
tags:
  - name: menu
    description: Operations about menu
schemes:
  - https
  - http
paths:
  /menu:
    post:
      tags:
        - menu
      summary: Create menu
      operationId: createMenu
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/Menu'
      responses:
        default:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Menu'
    get:
      tags:
        - menu
      summary: List all menu
      operationId: listMenu
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Menu'
        '400':
          description: Invalid menu id supplied
        '404':
          description: Menu not found
  '/menu/{id}':
    get:
      tags:
        - menu
      summary: Get menu by menu id
      operationId: readMenu
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Menu'
        '400':
          description: Invalid menu id supplied
        '404':
          description: Menu not found
    put:
      tags:
        - menu
      summary: Updated menu
      operationId: updateMenu
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - in: body
          name: body
          description: Updated menu object
          required: true
          schema:
            $ref: '#/definitions/Menu'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Menu'
        '400':
          description: Invalid menu supplied
        '404':
          description: Menu not found
    delete:
      tags:
        - menu
      summary: Delete menu
      operationId: deleteMenu
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '400':
          description: Invalid menu id supplied
        '404':
          description: Menu not found
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: implicit
    scopes:
      'write:pets': modify pets in your account
      'read:pets': read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  Menu:
    type: object
    properties:
      id:
        type: integer
        format: int32
      name:
        type: string
      content:
        type: string
      created_by:
        type: string
      created_at:
        type: integer
        format: int64
      updated_by:
        type: string
      updated_at:
        type: integer
        format: int64
    xml:
      name: Menu